+++ source/ftrack_connect/ui/widget/web_view.py
@@ -1,7 +1,18 @@
 # :coding: utf-8
 # :copyright: Copyright (c) 2015 ftrack
 
-from QtExt import QtGui, QtWebKitWidgets, QtCore, QtWidgets
+from QtExt import QtGui, QtCore, QtWidgets, QtNetwork
+
+try:
+    from QtExt import QtWebKitWidgets as QtWebWidgets
+    HAS_WEBKIT=True
+except ImportError:
+    from QtExt import QtWebEngineWidgets as QtWebWidgets
+    HAS_WEBKIT=False
+    # Create some aliases for old QtWebKit classes.
+    QtWebWidgets.QWebPage = QtWebWidgets.QWebEnginePage
+    QtWebWidgets.QWebView = QtWebWidgets.QWebEngineView
+
 
 # TODO: Investigate why this import exists and remove it.
 try:
@@ -24,7 +35,7 @@ class Ui_WebView(object):
         self.horizontalLayout = QtWidgets.QHBoxLayout(WebView)
         self.horizontalLayout.setContentsMargins(0, 0, 0, 0)
         self.horizontalLayout.setObjectName('horizontalLayout')
-        self.WebViewView = QtWebKitWidgets.QWebView(WebView)
+        self.WebViewView = QtWebWidgets.QWebView(WebView)
         font = QtGui.QFont()
         font.setFamily('Anonymous Pro')
         self.WebViewView.setFont(font)
@@ -44,14 +55,6 @@ class Ui_WebView(object):
         )
 
 
-class WebPage(QtWebKitWidgets.QWebPage):
-    '''WebPage widget.'''
-
-    def javaScriptConsoleMessage(self, msg, line, source):
-        '''Print javascript console message.'''
-        print '%s line %d: %s' % (source, line, msg)
-
-
 # TODO: Remove this widget and refactor Maya plugin to use WebView.
 class WebViewWidget(QtWidgets.QWidget):
     '''Webview widget class.'''
@@ -62,13 +65,20 @@ class WebViewWidget(QtWidgets.QWidget):
         self.ui = Ui_WebView()
         self.ui.setupUi(self)
 
-        self.webPage = WebPage()
+        self.webPage = QtWebWidgets.QWebPage()
+
         self.persCookieJar = PersistentCookieJar(self)
         self.persCookieJar.load()
 
         proxy = HelpFunctions.getFtrackQNetworkProxy()
         if proxy:
-            self.webPage.networkAccessManager().setProxy(proxy)
+            if HAS_WEBKIT:
+                self.webPage.networkAccessManager().setProxy(proxy)
+            else:
+                # TODO: This is a global per QApplication setting.
+                # Changing it might not be a good idea?
+                # QtNetwork.QNetworkProxy.setApplicationProxy(proxy)
+                pass
 
         self.ui.WebViewView.setPage(self.webPage)
 
@@ -100,7 +110,7 @@ class WebView(QtWidgets.QFrame):
         layout.setSpacing(0)
         self.setLayout(layout)
 
-        self._webView = QtWebKitWidgets.QWebView()
+        self._webView = QtWebWidgets.QWebView()
         layout.addWidget(self._webView)
 
         self.set_url(url)
@@ -116,3 +126,14 @@ class WebView(QtWidgets.QFrame):
             return None
 
         return url
+
+    def evaluateJavascript(self, javascript):
+        '''Evaluate a Javascript script on the webpage's main frame.
+        This method was moved from the QWebFrame class in QtWebkit
+        to the QWebPage class in QtWebEngine.
+        We wrap it here for convenience.
+        '''
+        if HAS_WEBKIT:
+            self._webView.page().mainFrame().evaluateJavaScript(javascript)
+        else:
+            self._webView.page().evaluateJavaScript(javascript)