+++ doc/release/release_notes.rst
@@ -8,7 +8,22 @@
 Release Notes
 *************
 
-.. release:: Upcoming
+ .. release:: Upcoming
+    .. change:: fixed
+        :tags: Ui
+
+        Info and Tasks dialogs are now compatible with Qt5, PySide2 and QtWebEngine.
+
+.. release:: 1.0.1
+    :date: 2017-07-11
+
+    .. change:: fixed
+        :tags: Asset manager
+
+        Cannot change version of versions with a sequence component.
+
+.. release:: 1.0.0
+    :date: 2017-07-07
 
     .. change:: fixed
         :tags: API